requirements.txt
python-docx>=1.1.0
lxml>=4.9.0
diff-match-patch>=20230430
docx2pdf>=0.10
PyMuPDF>=1.23.0
